refactor(mcp): own JSON-RPC connection scheduling
