+++ crates/tracedecay-mcp/Cargo.toml
@@ -42,7 +42,7 @@ schemars = "1.2.1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 thiserror = "2"
-tokio = { version = "1", features = ["rt", "macros", "io-util", "io-std", "sync", "time"] }
+tokio = { version = "1", features = ["rt", "macros", "io-util", "io-std", "signal", "sync", "time"] }
 toml = "1"
 tracedecay-api = { path = "../tracedecay-api", version = "0.1.0" }
 tracedecay-contracts = { path = "../tracedecay-contracts", version = "0.1.0" }
