+++ crates/tracedecay/src/mcp/server/routing.rs
@@ -4,6 +4,7 @@
 use std::path::{Path, PathBuf};
 
 use serde_json::Value;
+use tracedecay_mcp::server::{McpConnectionState, McpResponseLease};
 
 use crate::mcp::project_route::{
     HookProjectRouteCache, ProjectRouteFailure, ProjectRouteFailureKind, WorkspaceProjectRoute,
@@ -52,6 +53,12 @@ impl SelectedProjectResponseLease {
 
     pub(crate) fn is_revoked(&self) -> bool {
         self.revoked.is_cancelled()
+    }
+}
+
+impl McpResponseLease for SelectedProjectResponseLease {
+    fn revoked(&self) -> &tracedecay_session_memory::context::CancellationToken {
+        &self.revoked
     }
 }
 
@@ -158,6 +165,22 @@ impl ConnectionRouteState {
 
     pub(crate) fn clear_selected_request_server(&mut self) {
         self.selected_request_server = None;
+    }
+}
+
+impl McpConnectionState for ConnectionRouteState {
+    type ResponseLease = SelectedProjectResponseLease;
+
+    fn memory_request_scope(&self) -> &str {
+        ConnectionRouteState::memory_request_scope(self)
+    }
+
+    fn fork_for_connection_owned_read(&self) -> Self {
+        ConnectionRouteState::fork_for_connection_owned_read(self)
+    }
+
+    fn take_selected_response_lease(&mut self) -> Option<Self::ResponseLease> {
+        ConnectionRouteState::take_selected_response_lease(self)
     }
 }
